+++ packages/patternfly-3/patternfly-react-extensions/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.16.1"></a>
+## [2.16.1](https://github.com/patternfly/patternfly-react/compare/patternfly-react-extensions@2.16.0...patternfly-react-extensions@2.16.1) (2019-01-03)
+
+
+
+
+**Note:** Version bump only for package patternfly-react-extensions
+
 <a name="2.16.0"></a>
 # [2.16.0](https://github.com/patternfly/patternfly-react/compare/patternfly-react-extensions@2.14.2...patternfly-react-extensions@2.16.0) (2018-12-19)